requirements.txt
attrs
cython
numpy==1.24.0
decorator
sympy
cffi
pyyaml
pathlib2
psutil
protobuf==3.20
scipy
requests
absl-py
cloudpickle
ml-dtypes
tornado
setuptools 
torch==2.6.0 --index-url https://download.pytorch.org/whl/cpu
torch-npu==2.6.0 
torch-tb-profiler-ascend
pybind11
expecttest
